fix node 6 support part 2
