+++ test/hooks.on-ready.test.js
@@ -3,7 +3,15 @@
 const semver = require('semver')
 const t = require('tap')
 const Fastify = require('../fastify')
-const immediate = require('util').promisify(setImmediate)
+
+if (semver.gt(process.versions.node, '8.0.0')) {
+  require('./hooks.on-ready.async')(t)
+} else {
+  t.test('async tests', t => {
+    t.plan(1)
+    t.pass('Skip because Node version < 8')
+  })
+}
 
 t.test('onReady should be called in order', t => {
   t.plan(7)
@@ -26,7 +34,9 @@ t.test('onReady should be called in order', t => {
 
     childOne.register((childTwo, o, next) => {
       childTwo.addHook('onReady', function () {
-        return immediate()
+        return new Promise((resolve) => {
+          setTimeout(resolve, 0)
+        })
           .then(() => {
             t.equals(order++, 2, 'called in childTwo')
             t.equals(this.pluginName, childTwo.pluginName, 'the this binding is the right instance')
@@ -38,132 +48,6 @@ t.test('onReady should be called in order', t => {
   })
 
   fastify.ready(err => t.error(err))
-})
-
-t.test('async onReady should be called in order', { skip: semver.lt(process.versions.node, '8.0.0') }, async t => {
-  t.plan(7)
-  const fastify = Fastify()
-
-  let order = 0
-
-  fastify.addHook('onReady', async function () {
-    await immediate()
-    t.equals(order++, 0, 'called in root')
-    t.equals(this.pluginName, fastify.pluginName, 'the this binding is the right instance')
-  })
-
-  fastify.register(async (childOne, o) => {
-    childOne.addHook('onReady', async function () {
-      await immediate()
-      t.equals(order++, 1, 'called in childOne')
-      t.equals(this.pluginName, childOne.pluginName, 'the this binding is the right instance')
-    })
-
-    childOne.register(async (childTwo, o) => {
-      childTwo.addHook('onReady', async function () {
-        await immediate()
-        t.equals(order++, 2, 'called in childTwo')
-        t.equals(this.pluginName, childTwo.pluginName, 'the this binding is the right instance')
-      })
-    })
-  })
-
-  await fastify.ready()
-  t.pass('ready')
-})
-
-t.test('mix ready and onReady', async t => {
-  t.plan(2)
-  const fastify = Fastify()
-  let order = 0
-
-  fastify.addHook('onReady', async function () {
-    await immediate()
-    order++
-  })
-
-  await fastify.ready()
-  t.equals(order, 1)
-
-  await fastify.ready()
-  t.equals(order, 1, 'ready hooks execute once')
-})
-
-t.test('listen and onReady order', async t => {
-  t.plan(9)
-
-  const fastify = Fastify()
-  let order = 0
-
-  fastify.register((instance, opts, next) => {
-    instance.ready(checkOrder.bind(null, 0))
-    instance.addHook('onReady', checkOrder.bind(null, 4))
-
-    instance.register((subinstance, opts, next) => {
-      subinstance.ready(checkOrder.bind(null, 1))
-      subinstance.addHook('onReady', checkOrder.bind(null, 5))
-
-      subinstance.register((realSubInstance, opts, next) => {
-        realSubInstance.ready(checkOrder.bind(null, 2))
-        realSubInstance.addHook('onReady', checkOrder.bind(null, 6))
-        next()
-      })
-      next()
-    })
-    next()
-  })
-
-  fastify.addHook('onReady', checkOrder.bind(null, 3))
-
-  await fastify.ready()
-  t.pass('trigger the onReady')
-  await fastify.listen(0)
-  t.pass('do not trigger the onReady')
-
-  await fastify.close()
-
-  function checkOrder (shouldbe) {
-    t.equals(order, shouldbe)
-    order++
-  }
-})
-
-t.test('multiple ready calls', { skip: semver.lt(process.versions.node, '8.0.0') }, async t => {
-  t.plan(11)
-
-  const fastify = Fastify()
-  let order = 0
-
-  fastify.register(async (instance, opts) => {
-    instance.ready(checkOrder.bind(null, 1))
-    instance.addHook('onReady', checkOrder.bind(null, 6))
-
-    await instance.register(async (subinstance, opts) => {
-      subinstance.ready(checkOrder.bind(null, 2))
-      subinstance.addHook('onReady', checkOrder.bind(null, 7))
-    })
-
-    t.equals(order, 0, 'ready and hooks not triggered yet')
-    order++
-  })
-
-  fastify.addHook('onReady', checkOrder.bind(null, 3))
-  fastify.addHook('onReady', checkOrder.bind(null, 4))
-  fastify.addHook('onReady', checkOrder.bind(null, 5))
-
-  await fastify.ready()
-  t.pass('trigger the onReady')
-
-  await fastify.ready()
-  t.pass('do not trigger the onReady')
-
-  await fastify.ready()
-  t.pass('do not trigger the onReady')
-
-  function checkOrder (shouldbe) {
-    t.equals(order, shouldbe)
-    order++
-  }
 })
 
 t.test('onReady should manage error in sync', t => {
@@ -200,34 +84,6 @@ t.test('onReady should manage error in sync', t => {
   })
 })
 
-t.test('onReady should manage error in async', t => {
-  t.plan(4)
-  const fastify = Fastify()
-
-  fastify.addHook('onReady', function (done) {
-    t.pass('called in root')
-    done()
-  })
-
-  fastify.register(async (childOne, o) => {
-    childOne.addHook('onReady', async function () {
-      t.pass('called in childOne')
-      throw new Error('FAIL ON READY')
-    })
-
-    childOne.register(async (childTwo, o) => {
-      childTwo.addHook('onReady', async function () {
-        t.fail('should not be called')
-      })
-    })
-  })
-
-  fastify.ready(err => {
-    t.ok(err)
-    t.equals(err.message, 'FAIL ON READY')
-  })
-})
-
 t.test('onReady should manage sync error', t => {
   t.plan(4)
   const fastify = Fastify()
@@ -237,17 +93,20 @@ t.test('onReady should manage sync error', t => {
     done()
   })
 
-  fastify.register(async (childOne, o) => {
+  fastify.register((childOne, o, next) => {
     childOne.addHook('onReady', function (done) {
       t.pass('called in childOne')
       throw new Error('FAIL UNWANTED SYNC EXCEPTION')
     })
 
-    childOne.register(async (childTwo, o) => {
-      childTwo.addHook('onReady', async function () {
+    childOne.register((childTwo, o, next) => {
+      childTwo.addHook('onReady', function () {
         t.fail('should not be called')
       })
+      next()
     })
+
+    next()
   })
 
   fastify.ready(err => {
@@ -264,7 +123,7 @@ t.test('onReady can not add decorators or application hooks', t => {
     t.pass('called in root')
     fastify.decorate('test', () => {})
 
-    fastify.addHook('onReady', async function () {
+    fastify.addHook('onReady', function () {
       t.fail('it will be not called')
     })
     done()
@@ -294,20 +153,9 @@ t.test('onReady cannot add lifecycle hooks', t => {
   })
 
   fastify.addHook('onRequest', (request, reply, done) => {})
-  fastify.get('/', async () => 'hello')
+  fastify.get('/', (r, reply) => reply.send('hello'))
 
   fastify.ready((err) => { t.ok(err) })
-})
-
-t.test('onReady throw loading error', t => {
-  t.plan(1)
-  const fastify = Fastify()
-
-  try {
-    fastify.addHook('onReady', async function (done) {})
-  } catch (e) {
-    t.true(e.message === 'Async function has too many arguments. Async hooks should not use the \'done\' argument.')
-  }
 })
 
 t.test('onReady does not call done', t => {
